+++ firmware/controllers/can/can.h
@@ -45,13 +45,6 @@ void processCanRxMessage(const CANRxFrame& msg, efitick_t nowNt);
 void registerCanListener(CanListener& listener);
 void registerCanSensor(CanSensorBase& sensor);
 
-// Indicate that an ack response was received from the wideband bootloader
-void handleWidebandBootloaderAck();
-// Update the firmware on any connected wideband controller
-void updateWidebandFirmware();
-// Set the CAN index offset of any attached wideband controller
-void setWidebandOffset(uint8_t index);
-
 class CanWrite final : public PeriodicController<512> {
 public:
 	CanWrite();