+++ firmware/CHANGELOG.md
@@ -32,7 +32,10 @@ All notable user-facing or behavior-altering changes will be documented in this 
 
 ### Added
  - Injector nonlinearity (small pulse) correction - so far just polynomial, but table modes coming soon.
- - 1-4-3-6-2-5 firing order for VAG v6  
+ - 1-4-3-6-2-5 firing order for VAG v6
+ 
+### Fixed
+ - rusEFI console start-up speed #2964   
 
 ## June 2021 Release "National Logistics Day"